Жанры пофиксены
Процесс замены ID на названия интегрирован в функцию создания разметки
